Extracted alertExpiredFoods into its own class, added program help command, modified grammar in PerishableFood
